ai_powered_clinical_data_normalizer/ai_powered_clinical_data_normalizer/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymupdf==1.25.5
pytesseract==0.3.13
pillow==11.2.1
python-multipart==0.0.20
